new update

+++ bukkit/src/main/java/net/craftingstore/bukkit/events/DonationReceivedEvent.java
@@ -5,7 +5,6 @@ import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public class DonationReceivedEvent extends Event implements Cancellable {
@@ -19,7 +18,7 @@ public class DonationReceivedEvent extends Event implements Cancellable {
     @Deprecated
     private String packageName;
     @Deprecated
-    private BigDecimal packagePrice;
+    private int packagePrice;
     @Deprecated
     private int couponDiscount;
 
@@ -28,7 +27,7 @@ public class DonationReceivedEvent extends Event implements Cancellable {
     private boolean cancelled = false;
 
     @Deprecated
-    public DonationReceivedEvent(String command, String username, UUID uuid, String packageName, BigDecimal packagePrice, int couponDiscount) {
+    public DonationReceivedEvent(String command, String username, UUID uuid, String packageName, int packagePrice, int couponDiscount) {
         super(true);
         this.command = command;
         this.username = username;
@@ -74,7 +73,7 @@ public class DonationReceivedEvent extends Event implements Cancellable {
     }
 
     @Deprecated
-    public BigDecimal getPackagePrice() {
+    public int getPackagePrice() {
         return packagePrice;
     }
